Replace snapshot tests with our own tests (plus test utilities)
Snapshot tests are cool, but I think they may be better for regression-style tests we run against our storybook stories, to ensure that nothing unintentional has happened.

For our component tests, it may make more sense to have explicit tests which:
- Explain our intent
- Provide an opportunity to test-drive

Note that to accomplish this we have also added some basic test utils. They are not really necessary, and we could just use the functions directly on the React TestRenderer instance. However, I want to explore providing our own, thoughtful, conservative API.

We will certainly find out if it is a good/bad idea or not.