Add ".module.scss" to all stylesheet file names
This is unfortunate...

I chose to use react-scripts to build our react-styleguidist stuff, instead of maintaining our own Webpack configuration. That feels better to me because in my experience per-project webpack configs always go out of date, and are never updated. It is also a lot to maintain different configs for every project.

However, there is obviously a downside with that - we are at the whims of react-scripts. And react-scripts will not process CSS files as CSS modules unless they have ".module" in them.

Bummer. From one standpoint, though, this is a positive - now it is clear that these are intended to be used as CSS modules.