Simplify fill and color logic
The logic was pretty redundant: if there is no fill or color then use defaults. But then it was also complex: if there is one of fill or color then use exact provided values -- why not just make it more explicit?

In addition, fill none is basically an undefined value. We might as well just use the API and default one of the provided values. Same with color.

This should encourage usages of SVGs with intention.