Update Icon tests to use `@testing-library/react` patterns
I also took the opportunity to loop unroll tests. This is more compatible with the new testing patterns since it renders to the DOM.

I also took the opportunity to be more explicit in the tests and avoid asserting other states of the components which were unrelated to the actual test -- the default classes are already tested in the snapshot!