Set CustomFieldInputNumber input type as a permanent `number`
The custom field number type has a very specific database schema and only accepts integers. We will refactor the existing code to use the same kind of patterns. The native number type already has a really good coverage on validity -- we just need to see if we can hook it up to the React component well.