Implement new invalid state machine by leveraging native browser validation on input element.
This seems to be working well on Chrome, Firefox, and Safari. I will be checking IE Edge in a bit but since the new Edge will be on Chromium, it is a low risk of not working.

The implementation uses the native browser validations on the number input type. It even has builtin internationalization on the validation messages.

This commit is WIP because the tests are behaving weirdly. It seems like the input element is invalid when passing a legit number to the component. It is failing on step mismatch on the value `1`. I want to dig deeper but I figure I should have a commit since this is getting a bit unwiedly.

The followups are Edge testing, Jest DOM investigation, and test coverage review.