Bump Cypress image to utilize Node 8.16.0
If we want to upgrade Jest, we need to upgrade a few other dependencies and they require a newer version of Node. I am not sure if we need to bump to the newer LTS. I will try to minimize the jump. Bigmaven devops infrastructure is a major bottleneck here since we cannot easily upgrade the Node version there... and I want to minimize the friction between developer environments since we do not know how to develop with docker containers.