Update Jest configuration to reflect latest
We can now provide multiple files to include before running tests. Since our test set up was only importing another file, we just use it as part of the configuration. Very nice!

I believe we have another PR that adds to the setup JS file, but we can deal with it then.