Hoist significant variables
It is a bit more organized -- where we reduce the risk of sprinkling code everywhere.
It is easier to change in the future whenever the API changes.

Pretty cool? Let us see how this holds up. It might be worthwhile to encourage the API team to expose a meta endpoint so we can on-the-fly get the latest values.