Do not blow away user input when the Number component is invalid
This has been an annoying bug:

In a real browser, typing `12..` would set the `<input type="number">` as invalid.
Using `onChange` to keep track of the value would actually keep track of the empty string ("")
because a NaN in `<input type="number">` will clear its value as part of its API but not in the browser.
Instead of using `onChange`, we only parse the value of the number on the blur event. If we know the
input is invalid, we do not even attempt to parse the value (since it will be wrong).

I tried writing unit tests for this. It does not work because the `<input>` API does not validate the same way
as a user typing. I investigated `jsdom` implementations and even tried doing the same commands in the Chrome
dev tools. Nothing worked.

I tried writing integration tests for this with our Cypress setup. It does not work because Cypress uses
JavaScript under the hood.

In the time being, we should verify proper functionality by QAing the components with appropriate text cases.

I guess this is more "cons" of using the HTML5 number spec. But I hope we are through with most of the weirdness.