Modify the CustomFieldInputCurrency value API to only accept subunit values of the currency
This allows our component to align with the Mavenlink API. We normally are dealing with money through subunits. Let us also keep the component with similar behavior to simplify reasoning (and thus simplify code). This is a major change but thankfully we are in beta still.

For example: if we want to represent $13.37 then we will pass in 1337 as a value and a currency code of USD. Our component already knows how to manipulate the cents because we are using some ISO standard and we are pulling that specification into our codebase.