Prevent some key event defaults, provide native event for onClear callback
Our keyboard and accessibility tool interactions also have meaning to the browser, and we don't want these interactions to cause other behavior. In this case, the behavior is primarily changing scroll location.

Additionally, providing the React `SyntheticEvent` to the callback isn't useful, due to React clearing out all the properties after the event is consumed. We capture and provide the `nativeEvent` for a meaningful callback context.