Remove bad ARIA behavior on default ariaLabel for Icon
No ARIA is better than bad ARIA. We do not want to use SVG file names because those tend to not express intent -- just design. For example, `icon-caret-down` is not an excuse to call an SVG a caret down in a screen reader. The button which opens the drodown will have its own label like `Open dropdown`.

Bonus: fix test with no expectation. smh.