feature(Form): implement Form component
Migrated from bigmaven. Given a list of refs, it can determine whether it is dirty or invalid. This assumes each field implements a dirty flag -- otherwise, we assume it is dirty. I am not really sure how we want to handle it in the long-term.

It also does some submission data handling. See example.