chore(Form): backfill `dirty` ref API tests
I chose to store one-use variables to better indicate intent of duck typing. The native logic is not really used in our production code but it is nice to support it? It was not hard to implement but it may be hard to maintain. If so, we can just refactor these tests to use our MDS Input which would have the correct interface.