feature(Input): Input component updates its value when given new props
BREAKING CHANGE

The Input component no longer supports `defaultValue` prop. Instead, it only uses `value` but still leaves the `<input>` as uncontrolled. In order to upgrade, replace the `defaultValue` prop with the `value` prop. As a bonus, if you are passing down an empty `onChange`, it can now be deleted without causing an internal React warning.