refactor(custom-field-input-multiple-choice): hoist ids to make it easier to grok
In addition, removing a test which explicitly checks for aria properties. I am not convinced this is a worthwhile test. Instead, I think I want to rely on a good QA shake down and good engineering practices around using the latest WAI-ARIA patterns.

At some point, it might make sense to enhance the queries and/or matchers in our test env to assert expect user behavior (instead of attributes on elements).