feat(date): close calendar on date control blur
Looks like `onBlur` only works with other elements which are focusable. This means clicking on the calendar presentational container will be a false positive. This might be something that is fixed later in browsers or we might want to consider having a more well-defined interactive widget design -- like the container has `tabIndex="-1"` or something.