doc(date): implement a ref example
This introduces a helper component to automatically print the ref's content. It expects the targeted component to have some hook into its on change effects. Theoretically, it should always exists since the ref is the imperative method for tracking the state while the on change handler is the functional method for tracking the state.